Testing
None of the testing works, it cannot reference the namespaces correctly.
I am not sure how to fix this, but this is being set up like my other
group and it works for us. The main thing is the folders are not set up
as separate projects within the solution.